Pull progress to separate file
