+++ extensions/ql-vscode/src/databaseFetcher.ts
@@ -18,7 +18,7 @@ import { retry } from "@octokit/plugin-retry";
 
 import { DatabaseManager, DatabaseItem } from "./local-databases";
 import { showAndLogInformationMessage, tmpDir } from "./helpers";
-import { reportStreamProgress, ProgressCallback } from "./commandRunner";
+import { reportStreamProgress, ProgressCallback } from "./progress";
 import { extLogger } from "./common";
 import { getErrorMessage } from "./pure/helpers-pure";
 import {
