+++ extensions/ql-vscode/src/query-server/run-queries.ts
@@ -1,7 +1,7 @@
 import { join } from "path";
 import { CancellationToken } from "vscode";
 import * as cli from "../cli";
-import { ProgressCallback } from "../commandRunner";
+import { ProgressCallback } from "../progress";
 import { DatabaseItem } from "../local-databases";
 import {
   getOnDiskWorkspaceFolders,
