+++ extensions/ql-vscode/src/quick-query.ts
@@ -11,7 +11,7 @@ import {
   getQlPackForDbscheme,
   showBinaryChoiceDialog,
 } from "./helpers";
-import { ProgressCallback, UserCancellationException } from "./commandRunner";
+import { ProgressCallback, UserCancellationException } from "./progress";
 import { getErrorMessage } from "./pure/helpers-pure";
 import { FALLBACK_QLPACK_FILENAME, getQlPackPath } from "./pure/ql";
 import { App } from "./common/app";
