+++ extensions/ql-vscode/src/variant-analysis/export-results.ts
@@ -13,7 +13,7 @@ import {
   ProgressCallback,
   UserCancellationException,
   withProgress,
-} from "../commandRunner";
+} from "../progress";
 import { showInformationMessageWithAction } from "../helpers";
 import { extLogger } from "../common";
 import { createGist } from "./gh-api/gh-api-client";
